requirements.txt
pyyaml==6.0.1
python-dotenv==1.0.0
aiofiles==23.2.0
click==8.1.7
rich==13.7.0
psutil==5.9.8